services/ai-bridge/pyproject.toml
[project]
name = "voxmail-ai"
version = "0.1.0"
description = "VoxMail AI service"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "asyncpg>=0.30",
  "httpx>=0.27",
  "python-multipart>=0.0.20",
  "python-socketio>=5.11",
]

[project.optional-dependencies]
dev = [
  "pyright",
  "ruff",
  "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexamail_ai"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["src"]
